+++ cmake/Versions.cmake
@@ -1,3 +1,4 @@
 cmake_minimum_required(VERSION 3.20)
 
 set(GTEST_VERSION 1.17.0 CACHE STRING "")
+set(BENCHMARCK_VERSION 1.9.5 CACHE STRING "")